backend-services/services/financial-reconciliation/src/requirements.txt
web3==6.0.0
prometheus-client==0.17.1
aiohttp==3.8.4
asyncio==3.4.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
requests==2.31.0
